requirements.txt
asyncio==3.4.3
beautifulsoup4==4.12.3
aiohttp==3.10.9
pandas==2.2.3
python-dotenv==1.0.1
requests==2.32.3
elasticsearch==7.10.1
numpy==1.25.0
celery==5.4.0
rabbitmq==0.2.0
uvicorn~=0.31.1
fastapi~=0.115.0
pydantic~=2.9.2